+++ service/controller/control_plane.go
@@ -11,25 +11,35 @@ import (
 	"github.com/giantswarm/operatorkit/resource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/retryresource"
+	"github.com/giantswarm/tenantcluster/v2/pkg/tenantcluster"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	apiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
 
 	"github.com/giantswarm/cluster-operator/pkg/label"
 	"github.com/giantswarm/cluster-operator/pkg/project"
 	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
+	"github.com/giantswarm/cluster-operator/service/controller/resource/controlplanestatus"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/deleteinfrarefs"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/keepforinfrarefs"
+	"github.com/giantswarm/cluster-operator/service/controller/resource/tenantclients"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/updateinfrarefs"
+	"github.com/giantswarm/cluster-operator/service/internal/basedomain"
+	"github.com/giantswarm/cluster-operator/service/internal/nodecount"
 	"github.com/giantswarm/cluster-operator/service/internal/releaseversion"
 )
 
 // ControlPlaneConfig contains necessary dependencies and settings for the
 // ControlPlane controller implementation.
 type ControlPlaneConfig struct {
+	BaseDomain     basedomain.Interface
 	K8sClient      k8sclient.Interface
 	Logger         micrologger.Logger
+	NodeCount      nodecount.Interface
+	Tenant         tenantcluster.Interface
 	ReleaseVersion releaseversion.Interface
 
 	Provider string
@@ -87,6 +97,35 @@ func NewControlPlane(config ControlPlaneConfig) (*ControlPlane, error) {
 func newControlPlaneResources(config ControlPlaneConfig) ([]resource.Interface, error) {
 	var err error
 
+	var controlPlaneStatusResource resource.Interface
+	{
+		c := controlplanestatus.Config{
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+			NodeCount: config.NodeCount,
+		}
+
+		controlPlaneStatusResource, err = controlplanestatus.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var tenantClientsResource resource.Interface
+	{
+		c := tenantclients.Config{
+			BaseDomain:    config.BaseDomain,
+			Logger:        config.Logger,
+			Tenant:        config.Tenant,
+			ToClusterFunc: newG8sControlPlaneToClusterFunc(config.K8sClient),
+		}
+
+		tenantClientsResource, err = tenantclients.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	var deleteInfraRefsResource resource.Interface
 	{
 		c := deleteinfrarefs.Config{
@@ -135,8 +174,16 @@ func newControlPlaneResources(config ControlPlaneConfig) ([]resource.Interface, 
 	}
 
 	resources := []resource.Interface{
+		tenantClientsResource,
+		// Following resources manage CR status information. Note that
+		// keepForInfraRefsResource needs to run before
+		// controlPlaneStatusResource because keepForInfraRefsResource keeps
+		// finalizers where controlPlaneStatusResource does not.
 		deleteInfraRefsResource,
 		keepForInfraRefsResource,
+		controlPlaneStatusResource,
+
+		// Following resources manage resources in the control plane.
 		updateInfraRefsResource,
 	}
 
@@ -170,4 +217,27 @@ func toG8sControlPlaneObjRef(obj interface{}) (corev1.ObjectReference, error) {
 	}
 
 	return key.ObjRefFromG8sControlPlane(cr), nil
+}
+
+func newG8sControlPlaneToClusterFunc(k8sClient k8sclient.Interface) func(ctx context.Context, obj interface{}) (apiv1alpha2.Cluster, error) {
+	return func(ctx context.Context, obj interface{}) (apiv1alpha2.Cluster, error) {
+		cr := &apiv1alpha2.Cluster{}
+		{
+			cp, err := key.ToG8sControlPlane(obj)
+			if err != nil {
+				return apiv1alpha2.Cluster{}, microerror.Mask(err)
+			}
+
+			// Note that we cannot use a key function here because we do not need to
+			// fetch the Control Plane again. We need to lookup the Cluster CR based
+			// on the G8sControlPlane CR. This is why we use types.NamespacedName here
+			// explicitly.
+			err = k8sClient.CtrlClient().Get(ctx, types.NamespacedName{Name: key.ClusterID(&cp), Namespace: cp.Namespace}, cr)
+			if err != nil {
+				return apiv1alpha2.Cluster{}, microerror.Mask(err)
+			}
+		}
+
+		return *cr, nil
+	}
 }