+++ service/controller/machine_deployment.go
@@ -26,8 +26,8 @@ import (
 	"github.com/giantswarm/cluster-operator/service/controller/resource/machinedeploymentstatus"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/tenantclients"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/updateinfrarefs"
-	"github.com/giantswarm/cluster-operator/service/controller/resource/workercount"
 	"github.com/giantswarm/cluster-operator/service/internal/basedomain"
+	"github.com/giantswarm/cluster-operator/service/internal/nodecount"
 	"github.com/giantswarm/cluster-operator/service/internal/releaseversion"
 )
 
@@ -35,6 +35,7 @@ type MachineDeploymentConfig struct {
 	BaseDomain     basedomain.Interface
 	K8sClient      k8sclient.Interface
 	Logger         micrologger.Logger
+	NodeCount      nodecount.Interface
 	Tenant         tenantcluster.Interface
 	ReleaseVersion releaseversion.Interface
 
@@ -128,6 +129,7 @@ func newMachineDeploymentResources(config MachineDeploymentConfig) ([]resource.I
 		c := machinedeploymentstatus.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
+			NodeCount: config.NodeCount,
 		}
 
 		machineDeploymentStatusResource, err = machinedeploymentstatus.New(c)
@@ -168,24 +170,9 @@ func newMachineDeploymentResources(config MachineDeploymentConfig) ([]resource.I
 		}
 	}
 
-	var workerCountResource resource.Interface
-	{
-		c := workercount.Config{
-			Logger: config.Logger,
-
-			ToClusterFunc: newMachineDeploymentToClusterFunc(config.K8sClient),
-		}
-
-		workerCountResource, err = workercount.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	resources := []resource.Interface{
 		// Following resources manage controller context information.
 		tenantClientsResource,
-		workerCountResource,
 
 		// Following resources manage CR status information. Note that
 		// keepForInfraRefsResource needs to run before