+++ service/controller/resource/controlplanestatus/error.go
@@ -1,4 +1,4 @@
-package workercount
+package controlplanestatus
 
 import (
 	"github.com/giantswarm/microerror"