+++ service/internal/nodecount/error.go
@@ -1,0 +1,30 @@
+package nodecount
+
+import "github.com/giantswarm/microerror"
+
+var invalidConfigError = &microerror.Error{
+	Kind: "invalidConfigError",
+}
+
+// IsInvalidConfig asserts invalidConfigError.
+func IsInvalidConfig(err error) bool {
+	return microerror.Cause(err) == invalidConfigError
+}
+
+var notFoundError = &microerror.Error{
+	Kind: "notFoundError",
+}
+
+// IsNotFound asserts notFoundError.
+func IsNotFound(err error) bool {
+	return microerror.Cause(err) == notFoundError
+}
+
+var tenantClusterNotInitializedError = &microerror.Error{
+	Kind: "tentantClusterNotInitializedError",
+}
+
+// IsTenantClusterInitialized asserts tenantClusterNotInitializedError.
+func IsTenantClusterInitialized(err error) bool {
+	return microerror.Cause(err) == tenantClusterNotInitializedError
+}